+++ src/components/page/main/Banner.tsx
