+++ src/components/page/main/Gametype.tsx
