+++ src/components/page/main/ListContainer.tsx
@@ -1,8 +1,8 @@
 import get from '@/actions/meeting/listAction';
-import Gametype from '@/components/main/Gametype';
-import Location from '@/components/main/Location';
-import Sort from '@/components/main/Sort';
-import Status from '@/components/main/Status';
+import Gametype from '@/components/page/main/Gametype';
+import Location from '@/components/page/main/Location';
+import Sort from '@/components/page/main/Sort';
+import Status from '@/components/page/main/Status';
 import { DefaultSearchParams } from '@/types/get.type';
 import WriteButton from '@/ui/Button/WriteButton';
 import Search from '@/ui/form/Search';
