--- src/components/post/detail/Comment/atom/User.tsx
@@ -1,5 +1,0 @@
-export default function User() {
-  return (
-    <div className="size-8 rounded-full border bg-white bg-[url(/asset/image/user/user.png)] bg-[length:95%_95%] bg-center bg-no-repeat md:size-[73px]" />
-  );
-}
