--- src/components/post/detail/Detail/View.tsx
@@ -1,47 +1,0 @@
-'use client';
-
-import Color from '@tiptap/extension-color';
-import Highlight from '@tiptap/extension-highlight';
-import Link from '@tiptap/extension-link';
-import TextAlign from '@tiptap/extension-text-align';
-import TextStyle from '@tiptap/extension-text-style';
-import Underline from '@tiptap/extension-underline';
-import { EditorContent, useEditor } from '@tiptap/react';
-import StarterKit from '@tiptap/starter-kit';
-
-const View = ({ content }: { content: string }) => {
-  const editor = useEditor({
-    immediatelyRender: false,
-    shouldRerenderOnTransaction: false,
-    editable: false,
-    content,
-    extensions: [
-      StarterKit.configure({
-        heading: {
-          levels: [1, 2, 3],
-        },
-      }),
-      TextAlign.configure({
-        types: ['heading', 'paragraph'],
-      }),
-      Underline,
-      TextStyle,
-      Color,
-      Highlight.configure({ multicolor: true }),
-      Link.configure({
-        openOnClick: false,
-        autolink: true,
-        defaultProtocol: 'https',
-      }),
-    ],
-    editorProps: {
-      attributes: {
-        class:
-          'w-full prose prose-sm sm:prose-base p-5 focus:outline-none max-w-none box-border',
-      },
-    },
-  });
-
-  return <EditorContent editor={editor} />;
-};
-export default View;
