+++ src/provider/ModalProvider/ModalProvider.tsx
@@ -1,11 +1,16 @@
 'use client';
 
+/* 
+  @Todo
+  추후 modal을 한번에 관리하기 위해서 사용
+*/
 import {
   ModalContext,
   ModalDispatchContext,
 } from '@/provider/ModalProvider/ModalContext';
 import CustomModal from '@/ui/layout/CustomModal';
-import { ReactNode, useCallback, useMemo, useState } from 'react';
+import { usePathname } from 'next/navigation';
+import { ReactNode, useCallback, useEffect, useMemo, useState } from 'react';
 
 interface ModalState {
   Component: any;
@@ -14,6 +19,8 @@ interface ModalState {
 }
 
 const ModalProvider = ({ children }: { children: ReactNode }) => {
+  const pathname = usePathname();
+
   const [opendModals, setOpendModals] = useState<ModalState[]>([]);
 
   const openHandler = useCallback((Component: any, props: any) => {
@@ -36,6 +43,10 @@ const ModalProvider = ({ children }: { children: ReactNode }) => {
   const value = useMemo(() => {
     return { openHandler, closeHandler };
   }, [openHandler, closeHandler]);
+
+  useEffect(() => {
+    setOpendModals([]);
+  }, [pathname]);
 
   return (
     <ModalContext.Provider value={opendModals}>
