+++ src/ui/layout/header/atom/LoginBtn.tsx
@@ -1,12 +1,13 @@
 'use client';
 
-import Login from '@/components/auth/Login/Login';
-import useModal from '@/hooks/useModal';
+import LoginModal from '@/components/auth/Login/LoginModal';
+import ModalPortal from '@/components/modal/ModalPortal';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
+import { useState } from 'react';
 
 const LoginBtn = () => {
-  const { openHandler } = useModal();
+  const [isOpenModal, setIsOpenModal] = useState<boolean>(false);
 
   const { data: session } = useSession() as any;
   const router = useRouter();
@@ -15,17 +16,26 @@ const LoginBtn = () => {
     if (session && !session.accessToken) {
       return router.push('/signup');
     }
-    openHandler(Login, {});
+    setIsOpenModal(true);
+  };
+
+  const modalCloseHandler = () => {
+    setIsOpenModal(false);
   };
 
   return (
-    <button
-      type="button"
-      className="flex h-[35px] w-[74px] cursor-pointer items-center justify-center rounded bg-primary text-base font-semibold text-white transition-colors hover:bg-[#3f92e0]"
-      onClick={onLoginHandler}
-    >
-      로그인
-    </button>
+    <>
+      <button
+        type="button"
+        className="flex h-[35px] w-[74px] cursor-pointer items-center justify-center rounded bg-primary text-base font-semibold text-white transition-colors hover:bg-[#3f92e0]"
+        onClick={onLoginHandler}
+      >
+        로그인
+      </button>
+      <ModalPortal>
+        <LoginModal isOpen={isOpenModal} onClose={modalCloseHandler} />
+      </ModalPortal>
+    </>
   );
 };
